+++ src/main/java/com/example/demo/boundedContext/member/form/MemberModifyForm.java
@@ -1,0 +1,24 @@
+package com.example.demo.boundedContext.member.form;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class MemberModifyForm {
+
+    @NotBlank(message = "비밀번호를 입력해주세요.")
+    private String password;
+
+    @NotBlank(message = "비밀번호를 한 번 더 입력해주세요.")
+    private String password2;
+
+    @Email
+    private String email;
+
+    @NotBlank(message = "휴대폰 번호를 입력해주세요.")
+    private String phoneNumber;
+
+}
